Revert "Packed QuarterlyDistribution struct"
This reverts commit bb995cd4fc3933b21cdc022a8539dfd283aa4ef9.

+++ src/grants/GrantFund.sol
@@ -127,14 +127,12 @@ contract GrantFund is IGrantFund, ExtraordinaryFunding, StandardFunding {
             // set initial voting power and remaining voting power
             if (voter.votingPower == 0) {
 
-                uint128 newVotingPower = uint128(
-                    Maths.wpow(
-                        _getVotesSinceSnapshot(
-                            msg.sender,
-                            screeningStageEndBlock - VOTING_POWER_SNAPSHOT_DELAY,
-                            screeningStageEndBlock
-                        ), 2
-                    )
+                uint256 newVotingPower = Maths.wpow(
+                    _getVotesSinceSnapshot(
+                        msg.sender,
+                        screeningStageEndBlock - VOTING_POWER_SNAPSHOT_DELAY,
+                        screeningStageEndBlock
+                    ), 2
                 );
 
                 voter.votingPower          = newVotingPower;
@@ -249,14 +247,12 @@ contract GrantFund is IGrantFund, ExtraordinaryFunding, StandardFunding {
                 // set initial voting power and remaining voting power
                 if (voter.votingPower == 0) {
 
-                    uint128 newVotingPower = uint128(
-                        Maths.wpow(
-                            _getVotesSinceSnapshot(
-                                msg.sender,
-                                screeningStageEndBlock - VOTING_POWER_SNAPSHOT_DELAY,
-                                screeningStageEndBlock
-                            ), 2
-                        )
+                    uint256 newVotingPower = Maths.wpow(
+                        _getVotesSinceSnapshot(
+                            msg.sender,
+                            screeningStageEndBlock - VOTING_POWER_SNAPSHOT_DELAY,
+                            screeningStageEndBlock
+                        ), 2
                     );
 
                     voter.votingPower          = newVotingPower;
